pyproject.toml
[project]
name = "resume-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "streamlit (>=1.45.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "openai (>=1.77.0,<2.0.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit-authenticator (>=0.4.2,<0.5.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[tool.poetry]
packages = [{include = "resume_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
